requirements.txt
SQLAlchemy
SQLAlchemy-Utils
cffi
paramiko
docker
maxminddb-geolite2
mysqlclient
django
graphene-django
django-filter
django-graphql-jwt
pymongo
djongo
graphene-mongo